+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [v0.2.9] - 2024-04-11
+
 ### Added
 
 - Allow select an activity type for the Redmine tasks